# Don't initialize multiple default sidecars for ns