Currently, we initalize a default sidecar scope for each *service* -
but the scopes only differ by namespace. Instead, we can just create one
per namespace required.
```
name                              old time/op  new time/op
InitPushContext/gateways-8        5.90ms ± 0%  3.17ms ± 0%
InitPushContext/empty-8            268ns ± 0%   245ns ± 0%
InitPushContext/telemetry-8        273ns ± 0%   247ns ± 0%
InitPushContext/virtualservice-8   163ns ± 0%   149ns ± 0%
```
I often see this codepath using a large chunk of Pilot CPU in real world
tests, so I think the change is significant in these cases.

Benchmarking built on https://github.com/istio/istio/pull/25117



[ ] Configuration Infrastructure
[ ] Docs
[ ] Installation
[ ] Networking
[ ] Performance and Scalability
[ ] Policies and Telemetry
[ ] Security
[ ] Test and Release
[ ] User Experience
[ ] Developer Infrastructure