@howardjohn 2020-06-30
/retest

On Tue, Jun 30, 2020 at 9:49 AM Istio Automation <notifications@github.com>
wrote:

> @howardjohn <https://github.com/howardjohn>: The following test *failed*,
> say /retest to rerun all failed tests:
> Test name Commit Details Rerun command
> integ-security-k8s-tests_istio a2cec47
> <https://github.com/istio/istio/commit/a2cec4713f1587171a752ec61a028c878849cb1b>
> link
> <https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25118/integ-security-k8s-tests_istio/15297> /test
> integ-security-k8s-tests_istio
>
> Instructions for interacting with me using PR comments are available here
> <https://git.k8s.io/community/contributors/guide/pull-requests.md>. If
> you have questions or suggestions related to my behavior, please file an
> issue against the kubernetes/test-infra
> <https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:>
> repository. I understand the commands that are listed here
> <https://go.k8s.io/bot-commands>.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/istio/istio/pull/25118#issuecomment-651915847>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAEYGXKMT54QGE4NZPPSFRTRZIJQZANCNFSM4OMND65A>
> .
>