@istio-testing 2020-06-30
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-pilot-k8s-tests_istio | a2cec4713f1587171a752ec61a028c878849cb1b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25118/integ-pilot-k8s-tests_istio/15173) | `/test integ-pilot-k8s-tests_istio`
integ-multicluster-k8s-tests_istio | a2cec4713f1587171a752ec61a028c878849cb1b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25118/integ-multicluster-k8s-tests_istio/5631) | `/test integ-multicluster-k8s-tests_istio`
analyze-tests_istio | a2cec4713f1587171a752ec61a028c878849cb1b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25118/analyze-tests_istio/3) | `/test analyze-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->